main.cpp update- size_t

+++ main.cpp
@@ -8,7 +8,7 @@ string sentence;
 
 bool palindrome(string sentence)
 {
-    int x=0;
+    size_t x=0;
     sentence.erase(remove(sentence.begin(), sentence.end(), ' '), sentence.end());
     size_t length=sentence.length();
         if(length==0){cout<<"No sentence was given.";exit(0);}
